SDC-11536: DifferentTypesAsOffsetIT is failing
I feel that this is more a bug in the product - I changed a precondition
few patches back to verify that offset is always partitionable. However
this is not correct assumption since the origin does support
non-partition load and for that more types are supported.

Thus this patch basically reverts my changed state check.

Change-Id: I44003df54cbf722b7f5cb9ad20580832c92724f4