SDC-11714.  Enable Parallel Execution in JDBC Query Executor.
In this patch we walk through the incoming batch and
process the INSERTs first, then
process the UPDATES and other SQL second, and DELETES last.

Specifically for the second "batch" the UPDATEs,
we process them serially.

If we did not do this, there is a small chance
that there may be two UPDATEs "back-to-back" in the
incoming batch, and there is a tiny
chance we can process them out of order. Processing them
in order eliminates this risk.

In parallel mode, We commit after each statement,
on one hand due to the setting
of Auto Commit in the UI or if that's not set we explicitly
COMMIT after each statement finishes.  This is simply to ensure
there are no uncommited records which can have a dependent record
in another thread.

Also, note that with Autocommit unchecked,
performance is somewhat reduced due to the extra COMMIT statements
which are being generated by the JDBC Query Executor.

The number of database connections from the maximum pool size
UI element is used to control the number of parallel threads we use.

Maximum pool size and min idle should be set to the same value
as we want to work with a fixed-size thread pool.

Change-Id: I7a9694c96b98130b434cbde701b38e7b9b346947
Reviewed-on: https://review.streamsets.net/c/22305
Reviewed-by: Jeff Evans <jeff@streamsets.com>