SDC-11747: Disable installing multiple versions of an Enterprise Connector at the same time
This patch disables the ability for user to shoot themselves in the
foot. Since the UI operates on stage lib as a unique thing, the user
experience isn't the best - but it does seem 'good enough' for now and
for this edge case.

Change-Id: Ibace145d1e39784a76bb5f6dade113b73594e60e
(cherry picked from commit 820ab0d58cb95bffb58562ea0d1fe5863572dd4f)