SDC-11844: Snappy and LZO tests failing on HDP modules
It seems that we have copied both LZO as well as Snappy codecs long time
ago (2015!) and I don't think that we need to keep those these days. I
however don't want to remove them in so late SDC version, so I rather
took the approach of adding the missing new method calls.

Change-Id: Iffdaf8d4cc0ebe3a06e62810e4384b20969c733f
Reviewed-on: https://review.streamsets.net/c/22859
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>