SDC-11780: MapReduce Executor option to Change Stage Library after Upgrade
The reason why the ability to change stage library showed for some
things, but not all, is that the function that calculate possible stage
libraries was checking two things - the stage name as well as the stage
library. Thus if the library was upgraded in the new version, we have
not displayed the option to change the stage library.

I feel that such check is actually not necessary. The backend enforces
that all stages from all stage libraries are on the same version and
data collector will refuse to start if that is not the case. Thus I
removed the version check.

Change-Id: I30d08574fd3284172cc8989799333c801f5099e5
Reviewed-on: https://review.streamsets.net/c/22901
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
(cherry picked from commit db155da8e3b344d3f36280d1ab54a3b2819cbc9b)
Reviewed-on: https://review.streamsets.net/c/22902
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>