SDC-11909: Pre-load known JDBC drivers if they are available on the classpath
Sometimes we see that the combination of security manager, our classpath
isolation and Java services fails to properly register all available
JDBC drivers. It seems as some inherent flakiness in Java on certain (or
all) Java implementations.

This patch is trying to ease the pain by explicitly pre-loading known
JDBC drivers to make sure that they are available even if this flakiness
occurs. There is no need to have all the drivers on the classpath though
- the capability is written in non-intrusive mode where we just attempt
and not fail when the driver(s) are not present.

Change-Id: I6ccc869a47e8af6eab848999fd9f25452dae6dd2
Reviewed-on: https://review.streamsets.net/c/23082
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>