SDC-11911: Add ability to explicitly configure certain JDBC drivers to load on per-SDC instance
This is last patch in the queue for improving behavior of 'missing'
jdbc drivers. In case that both the other methods fails, this patch
enables an admin to configure list of JDBC drivers that must be always
loaded - e.g. one doesn't have to update all pipelines in the SDC, just
one small config in sdc.properties.

Change-Id: Id5bc0b0fa8cf1cfb5b37009a709d7d49e7f6c280
Reviewed-on: https://review.streamsets.net/c/23089
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>