SDC-12043: Antenna Doctor: Propagate messages on failed preview
The preview path is different then normal execution of a pipeline and
thus I needed to provide special wiring to ensure that the messages are
properly displayed. Since I was already changing this code, I also added
stacktrace propagation to the UI - which is something that was generally
missing in preview before.

Change-Id: Ic80cf774641cc17b062b83a1110e1c91e0e96b1f
Reviewed-on: https://review.streamsets.net/c/23418
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>