SDC-12120. Old Upgrade tests are failing
SelectorStageUpgrader was wrongly invoking the deprecated upgrade()
method when running the (legacy) stage upgrader. This is a problem for
stages which directly implement the new upgrade() interface method. In
those cases, calling the deprecated method makes no changes in the
stage config.

This patch replace the upgrader invocations in SelectoStageUpgrader to
use the new method.

Change-Id: I8d1075fd628fea7ec9d5bf93899aa27720fe3c24
Reviewed-on: https://review.streamsets.net/c/23773
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>