SDC-12225: Ensure that Runner.getPipelineConfig returns validated and upgraded version of the pipeline
We have added public getPipelineConfiguration() method specifically for
stats aggregation - however this method returns the raw version of the
pipeline as is on disk. This version of the pipeline might not be
properly upgraded - that is why the rest of the Runner code actually
works with slightly different method that ensures that the pipeline is
in-fact upgraded after read.

This patch switches the implementation to the other method that properly
upgrades the pipeline.

Change-Id: I144efaf1c9bdee0e8dc2d9e58390112414135a5a
Reviewed-on: https://review.streamsets.net/c/24752
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
(cherry picked from commit 21c743efe6abda04f6aaacad4497f31b91d122d0)
Reviewed-on: https://review.streamsets.net/c/25250
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>