SDC-12303. NPE in SQL Parser.
When using Oracle CDC followed in the pipeline by
SQL Parser there is a chance to get an NPE.

We were testing the combinations of Oracle CDC and SQL Parser.
We forgot to update the Oracle CDC Origin's config and left
the "Parse SQL" option checked.  This NPE occurs in
SQL Parser because the SQL statement has already been
parsed in the origin; the /sql field does not exist.

Change-Id: I0161d45550b698aa11b46d31cb5c28a7b501bca5
Reviewed-on: https://review.streamsets.net/c/25206
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>