SDC-12311.  Oracle CDC - reduce memory use of addRecordsToQueue()
Please see a related change to support enabling/disabline this code in
SDC-12454.

In Oracle CDC, there is a routine addRecordsToQueue() which gets
the records from the internal storage (buffer in memory, buffer on disk).

There are a couple of loops in the routine, each loop create a list
of the records after processing.  These objects are held until the
end of the routine.

This routine is also the throttle for the outbound records - destined
to go through the pipeline.  The throttle is putting the records
into a BlockingQueue, so we may spend some seconds in this routine.

In the case of an enormous transaction (think > 1mm records) this
leads to a lot of memory being held until we exit the routine.

As far as SQL parallel parsing, we process in groups of 2 * batch size.

Change-Id: I0506a7dde87a367c8176b34e405177a80981d1c7
Reviewed-on: https://review.streamsets.net/c/25245
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>