SDC-12344: Make StatsCollectorRunner non-intrusive
This patch makes sure that the stats generation is non-intrusive,
meaning that when an exception is generated by stats aggregation, the
pipeline will continue executing properly. This is a follow up on my
earlier change done as SDC-12225 - which had the side effect that
invalid pipelines (due to the stats aggregator) were propagating
exception all the way up to the client rather then only updating
pipeline run status.

Change-Id: I57779d7bafc871145358332fda48e42063c24dff
Reviewed-on: https://review.streamsets.net/c/25398
Reviewed-by: Xavier Baqués <xavi@streamsets.com>