SDC-12354 Skip field validation when there are expressions
In SDC-12211, field paths are statically validated, but did not
account for expressions. It is too hard to validate the field path
when you can't evaluate expressions, so if there is an expression
then we simply skip validation and allow runtime validation to
catch any issues. Since the primary motivation for SDC-12211 was to
help new users, and new users are not likely to use expressions in
field paths, this doesn't miss the overall goals of SDC-12211.

Testing:
* new unit test that failed before making fix

Change-Id: I691629ea080a8c94fef22fb45070fbb9a11eabba
Reviewed-on: https://review.streamsets.net/c/25419
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>