SDC-12417. Thycotic Credential Store uses seconds as milliseconds
Also refactors some logic that didn't make sense:
getSecret().getSecretField(...) never actually would throw an Exception when
there was a problem, it would return null.  The updated logic now properly deals
with this.  It'll also keep using the older cached value if there is a problem,
instead of returning null.

Change-Id: Ie3534c1d9bee0c428e7f5100f87cc66757d30889
Reviewed-on: https://review.streamsets.net/c/25610
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>