SDC-12419: Update UI facing select boxes for non-canvas stages on pipeline upgrade
Non-canvas stages (pipeline start and stop, error, stats and test stage)
have special handling in the main pipeline conf (shown on general tab)
where we encode a duplicate information from the StageConfiguration -
the selected stage library, stage name and stage version. On pipeline
upgrade we were not properly updating this encoded version which made
the UI display empty box when an upgrade happened (and version was
increased). This patch explicitly updates those UI facing fields.

The only impact is on that select box in the UI, rest of the
functionality fully worked and thus it's more or less a cosmetic
problem.

Change-Id: Ie2343bb17bad6c0edf4d5b368746391510f2757f
Reviewed-on: https://review.streamsets.net/c/25613
Reviewed-by: Xavier Baqués <xavi@streamsets.com>