SDC-12486. Avoid BigQuery.getTable(TableId) throw a NullPointerException when TableId contains an empty name
BigQuery.getTable(tableId) throws a NullPointerException when called with a tableId with an empty name (empty, not null). This is a bug in the Google client implementation. This workaround intends to prevent our software to crash in this case and to show a warning to the user instead while sending their record to error.

Change-Id: Ifbdc01242d54ed4ccb077431b8cb6d993b88a90d
Reviewed-on: https://review.streamsets.net/c/25791
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>