SDC-12404. Pipeline Stuck in Running State on JDBC MultiTable exception (Exception not propagated/Pipeline Hangs)
JdbcUtil attempted to call java.sql.Date.toInstant and java.sql.Time.toInstant which are overriden to throw an UnsupportedOperationException. This is due to the fact that, in standard SQL, dates and times are different fields, meaning that a date is just a date, with no time, and times are just times, with no date.

This fix just adds the default missing Epoch values to dates and times. All dates will have "00:00:00" as their time, and all times will have "1970-01-01" as their date.

Change-Id: Ie26d251306fdedd550486c3999993ac86f5e7c2a
Reviewed-on: https://review.streamsets.net/c/25910
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>