SDC-12642. Improve ADLS origins performance when scanning directories
The previous implementation divided the task of scanning directories in two: first collect a list of
directories, and then for each directory list and filter files inside. This was done by using Hadoop
globStatus and getStatus calls, which is suboptimal. This patch reimplements this process by using the Hadoop
listFiles function, allowing to (when configured in stage) traverse recursivelly the directory tree.

Change-Id: I879bfb24693e343caa2fe14205a501db0a92da1a
Reviewed-on: https://review.streamsets.net/c/datacollector/+/26584
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hugo Prol <hugo@streamsets.com>