SDC-12740. Improve ADLS performance on populating pending files queue
ADLS origin stages use a queue for pending files which is sorted according to
the configured criterion (lexycographically or by modification timestamp). When
sorting by timestamp, the stages did not store the associated metadata. Instead,
they request that info to ADLS whenever is needed. This drastically slows down
the whole process. This patch modifies ADLS origins to store and reuse the
metadata to manage the pending files queue, without additional requests to ADLS
service.

Change-Id: I314e4429e4833e915c9731757acad47fe5f5897b
Reviewed-on: https://review.streamsets.net/c/datacollector/+/27117
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>