SDC-12744. Classification Previews don't always get cleaned up
Changes the expiration for the preview cache from 30 minutes to 5 - there's no
need to keep them around for that long.  Also adds a background thread to call
the cache cleanup method every 5 minutes because Guava isn't actually removing
expired previews otherwise.  We normally invalidate specific previews from the
cache when retrieving their output, so this isn't typically a problem; but if
the output isn't retrieved for some reason, the preview isn't cleaned up and
removed otherwise.

Change-Id: Ib8eee76c7693b480410ed66f8314b6130e7bf6ea
Reviewed-on: https://review.streamsets.net/c/datacollector/+/27176
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Pasindu Perera <pasindu@streamsets.com>
Reviewed-by: Tony McKeehan <anthony@streamsets.com>