SDC-12746. SFTP/FTP/FTPS Origin cannot handle rare case where offset is "-1"
In rare cases, if delegate#createOffset throws an exception when the origin is
processing the first file (e.g. sftp can't stat the file), it can cause the
offset to be "-1", and it subsequently can't handle that offset when #produce is
called again.  This patch updates #produce to properly handle a "-1" offset.
The new unit test forces this situation.

Change-Id: Id0e810d19ce61665ea04afac5c06a0dee3662ff0
Reviewed-on: https://review.streamsets.net/c/datacollector/+/27428
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>