SDC-12836. Type mismatch in return value when reading a Kudu Binary Column
Modified the functionality in KuduUtils when retrieving a Binary column from a
Kudu table. We attempt to create a Field and expect a ByteArray return type,
however Kudu's `RowResult.getBinary()` returns a ByteBuffer that wraps an array.
A conversion from ByteBuffer to ByteArray is required before hand to prevent
`KUDU_35 - Missing a value for column` exceptions.

Change-Id: I49971b966cdc3a7fce15e61097c2eb04b743272f
Reviewed-on: https://review.streamsets.net/c/datacollector/+/27429
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>