SDC-8738. Kafka destination pipeline does not retry upon Kafka stage exception when 'On Record Error: Stop Pipeline' is set at Kafka destination
StageException should not cause the pipeline to stop retrying just because the "On Record Error" action is set to "Stop Pipeline".
Remove the logic for checking calculateShouldStopOnStageError() altogether from the pipeline retry logic check

For e.g the Kakfa Producer destination could produce a StageException because of network Connectivity. This does not translate to an record error.
In this case, it is desired/expected that the pipeline should try to retry on failure (retry needs to be set in the configuration), as opposed to stopping pipeline

In addition, make sure to handle the following scenario:
When the "On Error Record : Stop Pipeline' is set, and if there is a valid error record situation, then stop the pipeline and do not retry EVEN IF retry pipeline is set to -1.

In Summary, expected behavior is as follows:
	- 	If pipeline is stopped by a StageException, and if retry is set, then retry starting the pipleine.
	- 	If the pipeline is stopped by an onErrorRecordException, and if error record action is to stop pipeline, then ignore whether or not retry is set.
		Simply stop pipeline and do not restart it

Change-Id: Iab8fea82805d96a37e3428b8580334c220241c76
Reviewed-on: https://review.streamsets.net/c/datacollector/+/27480
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>