SDC-13000. Solr destination fails to upgrade pipelines with SDC <= 3.1.0.0
When a pipeline is exported from version 3.1.1.0 to 3.12.0, it is missing ignoreOptionalFields and other parameters that were newly added between SDC-3.1.0.0 and SDC-3.12.
The current upgrader logic sets the "ignoreOptionalFields = FALSE" when doing pipeline upgrade from 3.1.0.0 to SDC-3.12
This can cause failures in existing pipelines in 3.1.0.0, if the Optional FieldMappings like "__root__, __version__, __text__" are not specified, when the pipeline is ported to SDC-3.12

Morevover in SolrConfigBean.Java, the default value for *ignoreOptionalFields* is "true"

So, the upgrader logic should also set "ignoreOptionalFields = true" to be self-consistent.

Change-Id: I16b599104d913a4cf5247a5cc0b20741214b5dee
Reviewed-on: https://review.streamsets.net/c/datacollector/+/28075
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>
(cherry picked from commit 218a67b5a31cb4256cd67e4c65a30c9c9ea9049e)
Reviewed-on: https://review.streamsets.net/c/datacollector/+/28047
Reviewed-by: Keith Burns <keith@streamsets.com>