SDC-12513: Amazon S3 origin - issue with converting string constent to null
This one is actually on me. The support for parsing nulls have to parts
- boolean (checkbox) controlling whether the replacement should happen
and then a secondary String configuration with a constant that is
considered null string.

Since the data format library have several layers, not all layers were
implemented properly. As it happens the lowest layer that actually does
the comparison was ignoring the boolean flag and rather directly
checking whether the string constant was present.

Thus what was happening is that user in the UI have checked the "Parse
Null" checkbox, filled the constant, then changed their mind and
unchecked the "Parse Null" checkbox - but this "unchecking" was pretty
much ignored since the constant was still present and used.

Change-Id: I9e5aa1faa2dcf980d93c979685cc7476d7dfab68
Reviewed-on: https://review.streamsets.net/c/datacollector/+/28137
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>