SDC-13835: Ship PostgreSQL and SQL Server JDBC Drivers with relevant stage libraries
This is actually technically a regression as in Data Collector we
intentionally ship SQL Server and PostgreSQL JDBC drivers with the
product as their license allows us to do that. It seems that we have
ported code that marked those dependencies as provided in jdbc-protolib.

I actually feel that having them provided in the jdbc-protolib is right
thing to do now when we have the branded libs as each stage lib should
define what driver(s) it should ship with. Which is exactly what this
patch does.

Change-Id: Ifedfcbcb5d9d1503346a6fdbd55cad5c49d05f70
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30278
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>