SDC-13813: Invalid decimal definition will end up with NullPointerException in our Avro Parser
It seems that there is no way we can do "ideal" case where we validate
schema before use in all cases. Especially since sometimes the schema
might be embedded in the file we're reading that Avro libraries will
fully handle on their own. Thus this patch improves the behavior only
slightly - rather then failing with NPE we clearly report which property
is missing.

Change-Id: Ifb710a526c79c4227ea02bdfeacb2f6234e00dfd
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30283
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>