SDC-13868. RJson Serializer Modifier overrode the Metric Module Serializers.
SDC-13598 introduced the RJson Mappers and SDC-13611 configured the global mapper with RJson customizations, but Rjson Customizations overrode the serializers for MetricsModule, causing the timers/histograms to misbehave. This has had impact on how metrics of a pipeline are rendered.
This patch fixes that by adding the serializer modifier rather than specifying the modifer to be used by the factory always.

Testing:
1. Unit test: New one added to make sure metrics serializers are not messed, all existing tests are successful
2. Tested new rest end points using RJson to make sure those still work.

Change-Id: Iaa6d8a95dd262fe1cc7988c44c1991569202861c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30330
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>