SDC-13867. Spark Executor does not generate appID field in event
SparkExecutor currently does not wait for the application to change state to submitted or beyond.
Which is presumably why appID is not generated/available yet from the spark handle

Add a config option called submitTimeout to wait for application to be submitted to spark.
Wait for spark job state to change to submitted until submitTimeout interval before getting appID
Set default = 0, i.e do not wait for submission

Change-Id: I6abd98ecd9ddb72e548810ffe4ee844e8158b04d
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30284
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>