SDC-13819: Kafka Multi-Topic Consumer refuses to ingest malformed records, rather than sending them to pipeline error handling
The problem is that our Kafka origin was sending a batch to the
framework for processing only if it contained at least 1 record. I
presume to avoid sending empty batches around. So in case that the batch
had no normal records and only error records, they were not properly
passed down to the pipeline.

Change-Id: I3901c786d64b90bd8969d7d93debe9de9b881fef
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30382
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Karthik Iyer <karthik@streamsets.com>