SDC-13900. PaginationInfoBinder randomly throws IllegalStateException.
This patch fixes the case where PaginationInfoInjectBinder randomly throws IllegalStateException on TestSecretResource. (which is 100% repeatable on Jenkins).
To fix that issue, instead of having the non final local PaginationInfo, we have a final member variable of HttpServletRequest and provide dynamically creates the pagination info.
This seem to have fixed the issue and jenkins seem to be happy with it.

Change-Id: Ie2660ff0f3cc032956e43ea20e9e24743ecac85e
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30394
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>