SDC-11777: provide way to easily see where a pipeline is when it is stuck in STARTING
In the spirit of what was suggested on the JIRA, this patch updates the
pipe runner to show which stage is currently running the init or destroy
phase.

It won't provide 100% visibility into what is happening during
init/destroy since there are plenty of activities happening outside of
stages on the canvas (lifecycle events, stats, error stage, ...), but
it's a good first step that doesn't interfere in case we decide to
overhaul the whole thing in the future.

Change-Id: I72f845a57181f40aca88ab7d92c4c805f7fae298
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30565
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>