SDC-13999. Cluster pipelines can fail to start due to non-existing SdcId.
This patch fixes the case where cluster pipelines fails to start because the SDCId is null/not present.This is because the unique sdcId for the slave pipelines are set after init tasks are executed in EmbeddedDataCollector. This patch makes sure the runTimeInfo is filled with the id for the particular slave before any of the initTask is called.

Testing:
1. Ran  stage/test_hadoop_fs_stages.py::test_hadoop_fs_origin_simple which was failing without the patch but passed after the patch.

Change-Id: Iaa1dc0bf0773bece1a60b18b18ff135e16d76fe9
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30626
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>