SDC-13751. Oracle CDC: Change LogMiner initialization to support Oracle 19c
* Modify LogMinerSession class to support sessions with CONTINUOUS_MINE disabled. When
CONTINUOUS_MINE is not active, now LogMinerSession manually manages the list of redo log
files required to mine the database for a given time range. This implies also to include
the redo logs containing a valid dictionary (if DICT_FROM_REDO_LOGS is configured) and any
other redo log file required to update the dictionary after any DDL change (if
DDL_DICT_TRACKING is enabled).

* Use CONTINUOUS_MINE option if the target database supports it (i.e. Oracle version <
19).

* Adapt the stage behavior to avoid LogMiner session windows to span into time points in
the future. In particular, when CONTINUOUS_MINE is enabled, any query to LogMiner whose
time range spans into the future has a blocking behavior: when reaching the last current
record, the query cursor blocks and waits for new database changes to be registered, until
the time range has been completed. However this behavior is not present when
CONTINUOUS_MINE is disabled; instead, when reaching the last current row, the query is
closed as usual in a database table. We avoid dealing with two different behaviors by
avoiding to span LogMiner session windows into the future.

Change-Id: I756fef295a1c3e91516136fcce0c62f9cc90f20f
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31109
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>