SDC-14033: MapR Stream tests are failing on MapR 5.2 and before
This is much larger patch then the required change because I wanted to
add unit tests (and that required a small refactoring).

But back to the problem, our FactoriesBean is validating that only one
'implementation' is provided on the classpath. But as it happens, we do
have a dependency chain on the implementations (e.g. Kafka 10 inheriting
from Kafka 9, ...). So to workaround that we have introduced workaround
where we explicitly configure allowed "duplicate" implementations and
which one should win. And this code wasn't updated to take MapR 5.2 into
account.

Change-Id: Ib35959ffa734ded223abe5ec744201ac0bb1ee21
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30720
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>