SDC-14050: test_kafka_origin_standalone.py::test_kafka_origin_timestamp_offset_strategy fails on secured cluster
The problem here is that the 0.11 Kafka consumer that is used as a base
for all Kafka consumers 0.11+ (including 1.0 and 2.0) have a concept of
auxiliary connection that isn't properly setting up the user provided
properties. Since kerberos is configured there, one can't ignore them
and must include them in the auxiliary connection too.

Change-Id: Ic0ffd3f109b2c8436094bd6a4bfc2115df51baf9
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30778
Reviewed-by: Karthik Iyer <karthik@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>