SDC-14184 Fix telemetry to save and roll with correct frequency
Save is supposed to happen every 60s, and on shutdown.
Roll is supposed to happen every 1 hour.
Reporting is supposed to happen every 24 hours.

This fixes a bug where save and roll were only running on the daily
trigger and on shutdown. The thread is now scheduled every 60s, and
will roll or report as needed, then save (always).

Also fixes a bug where we would consider any change to extraInfo to
be a cause to roll. extraInfo can have things like peak JVM memory
usage, which change over time should not cause a roll, so filtering
out extraInfo for the "sys info has changed" condition of rolling.

Synchronizes the runnable in case there are races on teardown.

Testing:
* New unit tests
* Manual testing - set roll period to 1 minute, report period to 2
  minutes, and configured to permit snapshot builds and upload to
  test bucket. Saw uploads with 2 entries as expected.

Change-Id: If22b6891aab058514ef9858b3d04bfc3b764a9ef
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31065
Reviewed-by: Santhosh Kumar <Santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>