SDC-14253: User libs must support concept of extra libs
I was troubleshooting failures in STF run for our Oracle Enterprise
connector where for some reason the test wasn't able to find the Oracle
driver. It took me a while to realize that we do not allow extra dir for
the 'user' connectors. But since we document to install our enterprise
connectors to user libs and then to use extra lib to drop JDBC drivers,
this seems as a major oversight on our end.

Change-Id: I82227e01b8561fe59fc3b1114bee42b17195223b
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31183
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>