SDC-14262: Selector Upgrader always runs YAML upgrader from the begging
The selector upgrader hardcoded overriding the "from" version to
whatever was the minimal version yaml upgrader supported. Meaning that
the whole yaml upgrader was used in all cases. E.g. even if the stage
was on version 10 and needed only upgrade to 11, if the yaml file
supported upgrade from 8, it was run from 8 to 11. This effectively lead
to loss of configuration since some of the properties in the pipeline
could be overridden.

Change-Id: If791bc029ffe2d15e9e39a71a565b923b3f13a86
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31208
Reviewed-by: Santhosh Kumar <Santhosh@streamsets.com>
(cherry picked from commit 8a94a2130c6d5141b6e087096cd2d4aa7919f178)
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31188