SDC-14309: MapR 5: Drop MapR 5 Stage libraries
First patch in longer series cleaning up and removing MapR 5. I want to
submit different patches so that it's easier to see what and why has
happened.

This one is removing the user visible stage libraries as well as
removing the packaging notes about them. All remaining code is however
kept in place for now.

Change-Id: I6732e13a11a2b60b9a198b2635e96994f9bf51ba
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31255
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Reviewed-by: Karthik Iyer <karthik@streamsets.com>