SDC-14475: Reset StringBuffer upon acquisition in log parser
Basically the problem I'm troubleshooting is that a user is pushing
binary data into the log parser which of course isn't working properly
since log is meant to be string. The particular problem that this change
addresses is small and in the grand scheme of things inconsequential, it
just helps with consistency in troubleshooting.

We are reusing StringBuffers across different records. The internal
implementation of the string buffer, since it's mean to work with
strings, is using binary zero to mark 'unused' bytes. The problem is
when a current record inserts zero bytes inside without explicitly
changing the length which on reuse causes the buffer to be confused
about what is inside and where the expected string ends. By re-setting
the length base to 0, we pretty much tell "ignore the other zeros
inserted in previous run".

Change-Id: If78cb103486e15d6de8b71dbc73b09168f77f4bd
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31599
Reviewed-by: Karthik Iyer <karthik@streamsets.com>