SDC-14489: JDBC Multitable origin must escape string columns
The problem is simple - we take values of all offset columns and create
a single string out of them, something like:

col1=value1::col2=value2

We however also support string type for the offset column which can
contain some of the same characters (::, =) which then breaks our
parsing logic.

This patch makes sure that we escape and unescape the offset values
accordingly.

Change-Id: I230d329b1225ee3c02a80091e1af7cdf8d4f19fa
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31633
Reviewed-by: Karthik Iyer <karthik@streamsets.com>