SDC-14452. Improve Oracle CDC initialization when DICT_FROM_REDO_LOGS
When DICT_FROM_REDO_LOGS and CONTINUOUS_MINE are set, the previous logic tried
to load a valid dictionary by starting an initial LogMiner session with a time
window of (now - durationDictExtract, durationDictExtract), where the
durationDictExtract is defined by the user. If this time window does not
contained a valid dictionary, then the stage started a wider LogMiner session
spanning the whole list of archived redo logs, which can be very slow.

It is possible to simplify this logic, since the LogMiner metadata contains info
about the available dictionaries in the archived redo logs. Thus, given a
starting SCN to mine, the new logic retrieves the most recent dictionary prior
to that SCN and starts a LogMiner session (SCN_dict_begin, SCN) to load it.

Change-Id: Idc9eb14b32b8562644b06486e071a23c107a59a3
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31667
Reviewed-by: Toni Martinez <toni@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>