SDC-14525. Loading all pipelines fail when a single pipeline has inconsistent metadata in data dir
Related to ESC-788

The issue is that even if a single pipeline has inconsistent metadata (state file doesn't exist, but pipeline info file exists),
then SDC UI fails to load the pipelines page and along with that the thread sending metadata about pipelines to SCH also fails.

The change makes sure to not fail bulk fetch of pipelines when only
certain pipelines have this inconsistency.

Change-Id: I21babe5bed282fbded951452e060b55364828dfb
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31707
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>