SDC-14597: Perf: FullPipeBatch.moveLaneCopying is creating one extra copy
The method moveLaneCopying takes the incoming lane and does deep copy
for all ongoing lanes - but since the incoming records were already
cloned or were provided to the framework as-is, discarding the data in
the input lane is wasteful. The overhead is 'only one more copy', but
since the fork is usually for 2-3 lanes, that one copy is significant
percentage of the overhead.

Change-Id: Ie7c2e9764577eedc9b221f379063dcf4a02fa6bb
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31820
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Reviewed-by: Karthik Iyer <karthik@streamsets.com>