SDC-14601. Oracle CDC: LogMiner restart error after internally closing session
This fixes a regression bug introduced by SDC-13751. When reading the LogMiner
dictionary from the redo logs, we must ensure the LogMiner dictionary is loaded
each time we begin a new LogMiner session.

We can either close the previous session and then load again the dictionary, or
avoid to explicitly close the previous session to keep the current loaded
dictionary (the old session is implicitly closed when starting a new LogMiner
session, but the dictionary is kept). We opt for this second option. Thus, the
only case where we need to reload the dictionary is after a database
reconnection.

Change-Id: Ic570111be3e5f9a8db1fad19cbb985c7d0264007
Reviewed-on: https://review.streamsets.net/c/datacollector/+/31828
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>