SDC-13347. Oracle CDC not mining new tables
The stage only tracked tables that existed during the pipeline
initialization. The list of tables to track were built during the initialization
by querying the database for the existing tables that matched the table names
and patterns configured in the stage. This prevented the stage to track any new
table created in the future which satisfied the configuration, and also
generate events for them.

The stage now directly uses the table name configuration as filters when
querying LogMiner, instead of using a list of concrete tables collected at
pipeline initialization.

For those tables satisfying the filters but dropped in the past, the stage will
not generate events nor CDC records, since the current implementation requires
the tables to exist in the database to extract their schema.

Change-Id: Ia28d8af617b51484512acb8251a55e9bdd04cbd6
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32016
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>