SDC-14382. SDC to automatically run Connection Upgrader on-the-fly
The original format used for storing Connection ConfigBeans was mirroring the
bean itself, which won't allow for upgrades or even changes to the ConfigBean.
This patch changes the format to some json that matches stages, pipelines, etc
and the way they are handled by BeanHelper, which we also re-use parts of.  This
also allows us to include some necessary extra metadata (e.g. the version of the
connection defintion).

The upgrade yaml is run using the YamlStageUpgrader.  A side-effect of this is
that some fields and log messages may call the connection a "stage", though
that's all superficial - the YamlStageUpgrader doesn't really care if it's a
stage or not.

The ConnectionOverlayInjectorContext was added as essentially a wrapper around
the given ConfigInjector.Context, but replacing the configs with the ones read
from the blob store.  It also massages the config names to properly handle
ConfigBeans inside of the Connection ConfigBean (e.g. "subbean.prop1").

Change-Id: I9cdcf5c4fe4d843637cf6208d46effedf7fe891f
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32202
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>