SDC-14808: MapR stage libraries ship copy of kafka-clients jar file
Confluent brings it's own version of kafka-clients jar which is actually a problem. MapR stage
libraries do not have the full classpath defined in this file, we have extra script that user
must run to install all the important jars 'from the cluster'. MapR is shipping their own custom
version of kafka-clients that must be taken from the 'cluster' jars and not any version we
downloaded from the internet.

Change-Id: I32253022d726938a877465de15228673add1192c
Reviewed-on: https://review.streamsets.net/c/datacollector/+/32307
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>